move chunks to std/io
